reduce number of required goroutines in sequentialPipe
This is a rather small change in the implementation of a pipe.
Instead of
- one channel per pipe
- one goroutine that
  - listens on this pipe and
  - spawns an additional goroutine per tuple and receiver
we now have
- one channel per receiver of a pipe
- one goroutine per such channel that
  - listens on that channel
  - and calls Box.Process() or Sink.Write() for each tuple,
thereby preventing additional goroutines for each tuple.

In linear and join topologies, this leads to an equivalent behavior
than the previous implementation. However, in fork topologies there
is a small difference because a tuple will *not* be enqueued in the
"all receivers' queue" when *all receivers* have free capacity any
more, but it will be enqueued in "a receiver's queue" when
*this receiver* has free capacity (depending on the loop order of
receivers, though), so some tuples may be processed earlier than
before.