fix the InputName for UDSF
The situation `FROM x AS a, x AS b` is quite different from
the situation `FROM x('param') AS a, x('param') AS b` topology-wise,
as in the former case there is just one box "x", but in the latter
there are two boxes and two edges. I dealt with that case by making
the InputName unique in the UDSF input case.