make `file` command similar to copy&paste behavior
Before this change, the whole file contents have been sent to the
SensorBee server in a single statement, which made debugging
harder (the individual statements are not written to the console)
and also didn't allow  `file` statements in files which were
themselves loaded using `file`. This change processes the lines
of a file just as if they were entered one by one into the SensorBee
shell, in particular the individual statements are echoed and it
is possible to use `file` within files.

Note that this is a *client* feature. Scripts that are loaded
from a `sensorbee run` configuration file cannot use this directive.